requirements.txt
ftfy==6.1.1
h5py==3.7.0
numpy==1.19.5
Pillow==9.4.0
regex==2021.10.23
timm==0.6.12
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.3
